foodtracker/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiosmtplib==4.0.1
    # via -r requirements.in
aiosqlite==0.21.0
    # via -r requirements.in
alembic==1.16.1
    # via -r requirements.in
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asgiref==3.8.1
    # via -r requirements.in
asyncpg==0.30.0
    # via -r requirements.in
authlib==1.6.0
    # via -r requirements.in
bcrypt==3.2.2
    # via -r requirements.in
billiard==4.2.1
    # via celery
celery[redis]==5.5.3
    # via -r requirements.in
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via
    #   bcrypt
    #   cryptography
    #   gevent
click==8.2.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   uvicorn
cryptography==45.0.4
    # via
    #   authlib
    #   python-jose
deprecated==1.2.18
    # via limits
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via pydantic
fastapi==0.115.12
    # via -r requirements.in
gevent==25.5.1
    # via -r requirements.in
greenlet==3.2.3
    # via
    #   gevent
    #   sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via -r requirements.in
jinja2==3.1.6
    # via -r requirements.in
kombu[redis]==5.5.4
    # via celery
limits==5.4.0
    # via slowapi
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
packaging==25.0
    # via
    #   kombu
    #   limits
    #   pytest
passlib==1.7.4
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.51
    # via click-repl
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic[email]==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via -r requirements.in
pygments==2.19.1
    # via pytest
pytest==8.4.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-dotenv
pytest-asyncio==1.0.0
    # via -r requirements.in
pytest-dotenv==0.5.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   pytest-dotenv
    #   uvicorn
python-jose[cryptography]==3.5.0
    # via -r requirements.in
python-magic==0.4.27
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via uvicorn
redis==5.2.1
    # via kombu
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
slowapi==0.1.9
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via
    #   -r requirements.in
    #   alembic
starlette==0.46.2
    # via fastapi
typing-extensions==4.14.0
    # via
    #   aiosqlite
    #   alembic
    #   fastapi
    #   limits
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via kombu
uvicorn[standard]==0.34.3
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==1.1.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via uvicorn
wrapt==1.17.2
    # via deprecated
zope-event==5.0
    # via gevent
zope-interface==7.2
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
